chore:패키지 생성 및 의존성 설정
- 아래 기준으로 패키지 생성
- 비즈니스 로직과 데이터를 포함하는 model
- 전체 애플리케이션을 제어하는 controller
- view와 같이 사용자와 상호작용하는 기능, 난수 생성 등 기타 유틸을 포함하여 utils
