+++ src/main/java/kr/codesquad/Main.java
@@ -1,5 +1,7 @@
 package kr.codesquad;
 
+import kr.codesquad.controller.LadderGameController;
+
 public class Main {
     public static void main(String[] args) {
         LadderGameController.run();
