+++ src/main/java/kr/codesquad/utils/Renderer.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.utils;
 
 import java.util.ArrayList;
 import java.util.List;
